requirements.txt
transformers
python-dotenv
